% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iobed.labels-package.R
\docType{package}
\name{iobed.labels-package}
\alias{iobed.labels}
\alias{iobed.labels-package}
\title{iobed.labels: Labels for IOBED data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functionality and pipelines to develop the files and infrastructure to labeling iobed data.
}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@ubep.unipd.it} (\href{https://orcid.org/0000-0002-0520-7428}{ORCID})

Authors:
\itemize{
  \item Maria Vittoria Chiaruttini \email{mariavittoria.chiaruttini@ubep.unipd.it} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
}

Other contributors:
\itemize{
  \item Dario Gregori \email{dario.gregori@ubep.unipd.it} (\href{https://orcid.org/0000-0001-7906-0580}{ORCID}) [contributor]
}

}
\keyword{internal}
